refactor!: name `metal-llvm` same as `metal-codegen-wasm` (#64)
