+++ crates/metal-codegen-llvm/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "metal-llvm"
+name = "metal-codegen-llvm"
 version = "0.1.0"
 edition = "2021"
-description = "The Metal LLVM backend implementation."
+description = "Support for turning MIR into LLVM IR"
 
 [dependencies]
 metal-mir = { workspace = true }
